Remove doc site and expose Swagger UI
