--- docs/Gemfile
@@ -1,3 +1,0 @@
-source "https://rubygems.org"
-
-gem "govuk_tech_docs"
