--- docs/manifest.yml
@@ -1,7 +1,0 @@
----
-applications:
-- name: qualified-teachers-api-tech-docs
-  memory: 64M
-  instances: 2
-  buildpacks:
-    - staticfile_buildpack
