--- docs/openapi-merge.json
@@ -1,14 +1,0 @@
-{
-  "inputs": [
-    {
-      "inputFile": "./tmp/v1.json"
-    },
-    {
-      "inputFile": "./tmp/v2.json"
-    },
-    {
-      "inputFile": "./tmp/v3.json"
-    }
-  ],
-  "output": "./source/dqt-api.yaml"
-}
