--- docs/source/index.html.md.erb
@@ -1,6 +1,0 @@
----
-title: DQT API
-description: API documentation for the Qualified Teachers API
----
-
-api>
