--- docs/source/javascripts/application.js
@@ -1,1 +1,0 @@
-//= require govuk_tech_docs
