--- docs/source/stylesheets/print.css.scss
@@ -1,3 +1,0 @@
-$is-print: true;
-
-@import "govuk_tech_docs";
