--- docs/source/stylesheets/screen-old-ie.css.scss
@@ -1,4 +1,0 @@
-$is-ie: true;
-$ie-version: 8;
-
-@import "govuk_tech_docs";
