--- docs/source/stylesheets/screen.css.scss
@@ -1,1 +1,0 @@
-@import "govuk_tech_docs";
